Uncommon AutoSIMD PIV increment node.
If the loop increment node is commoned, then reducing the loop by
changing the value of loop stride can cause unwanted bug. This is
potentially true if the node is commoned for address calculation.
This commit uncommon the PIV increment node so that changing the
stride does not introduce side-effects on commoned nodes.

Signed-off-by: Mohammad Nazmul Alam <mohammad.nazmul.alam@ibm.com>